refactor(main): extract internal subtitle extraction from main.ts
